test: convert featureTests to ava/typescript
