+++ test/snapshots/exports.ts.md
@@ -1,0 +1,197 @@
+# Snapshot report for `test/exports.ts`
+
+The actual snapshot is saved in `exports.ts.snap`.
+
+Generated by [AVA](https://ava.li).
+
+## exports the correct objects
+
+> Snapshot 1
+
+    {
+      IChangeEventPayload: undefined,
+      IDefaultEventPayload: undefined,
+      IDictionary: undefined,
+      ILinkOptions: undefined,
+      IModelOptions: undefined,
+      IModelPropertyDefinition: undefined,
+      IModelPropertyDefinitions: undefined,
+      INohmPrefixes: undefined,
+      IRelationChangeEventPayload: undefined,
+      ISortOptions: undefined,
+      IStaticMethods: undefined,
+      LinkError: Function LinkError {},
+      Nohm: NohmClass {
+        LinkError: Function LinkError {},
+        ValidationError: Function ValidationError {},
+        extraValidators: [],
+        isPublishSubscribed: false,
+        meta: true,
+        modelCache: {},
+        prefix: {
+          channel: 'nohm:channel:',
+          hash: 'nohm:hash:',
+          idsets: 'nohm:idsets:',
+          incrementalIds: 'nohm:ids:',
+          index: 'nohm:index:',
+          meta: {
+            idGenerator: 'nohm:meta:idGenerator:',
+            properties: 'nohm:meta:properties:',
+            version: 'nohm:meta:version:',
+          },
+          relationKeys: 'nohm:relationKeys:',
+          relations: 'nohm:relations:',
+          scoredindex: 'nohm:scoredindex:',
+          unique: 'nohm:uniques:',
+        },
+        publish: false,
+      },
+      NohmClass: Function NohmClass {},
+      NohmModel: Function NohmModelExtendable {},
+      TLinkCallback: undefined,
+      TTypedDefinitions: undefined,
+      ValidationError: Function ValidationError {},
+      boolProperty: 'bool',
+      dateProperty: 'date',
+      default: NohmClass {
+        LinkError: Function LinkError {},
+        ValidationError: Function ValidationError {},
+        extraValidators: [],
+        isPublishSubscribed: false,
+        meta: true,
+        modelCache: {},
+        prefix: {
+          channel: 'nohm:channel:',
+          hash: 'nohm:hash:',
+          idsets: 'nohm:idsets:',
+          incrementalIds: 'nohm:ids:',
+          index: 'nohm:index:',
+          meta: {
+            idGenerator: 'nohm:meta:idGenerator:',
+            properties: 'nohm:meta:properties:',
+            version: 'nohm:meta:version:',
+          },
+          relationKeys: 'nohm:relationKeys:',
+          relations: 'nohm:relations:',
+          scoredindex: 'nohm:scoredindex:',
+          unique: 'nohm:uniques:',
+        },
+        publish: false,
+      },
+      floatProperty: 'float',
+      integerProperty: 'integer',
+      jsonProperty: 'json',
+      nohm: NohmClass {
+        LinkError: Function LinkError {},
+        ValidationError: Function ValidationError {},
+        extraValidators: [],
+        isPublishSubscribed: false,
+        meta: true,
+        modelCache: {},
+        prefix: {
+          channel: 'nohm:channel:',
+          hash: 'nohm:hash:',
+          idsets: 'nohm:idsets:',
+          incrementalIds: 'nohm:ids:',
+          index: 'nohm:index:',
+          meta: {
+            idGenerator: 'nohm:meta:idGenerator:',
+            properties: 'nohm:meta:properties:',
+            version: 'nohm:meta:version:',
+          },
+          relationKeys: 'nohm:relationKeys:',
+          relations: 'nohm:relations:',
+          scoredindex: 'nohm:scoredindex:',
+          unique: 'nohm:uniques:',
+        },
+        publish: false,
+      },
+      numberProperty: 'number',
+      stringProperty: 'string',
+      timeProperty: 'time',
+      timestampProperty: 'timestamp',
+    }
+
+> Snapshot 2
+
+    NohmClass {
+      LinkError: Function LinkError {},
+      ValidationError: Function ValidationError {},
+      extraValidators: [],
+      isPublishSubscribed: false,
+      meta: true,
+      modelCache: {},
+      prefix: {
+        channel: 'nohm:channel:',
+        hash: 'nohm:hash:',
+        idsets: 'nohm:idsets:',
+        incrementalIds: 'nohm:ids:',
+        index: 'nohm:index:',
+        meta: {
+          idGenerator: 'nohm:meta:idGenerator:',
+          properties: 'nohm:meta:properties:',
+          version: 'nohm:meta:version:',
+        },
+        relationKeys: 'nohm:relationKeys:',
+        relations: 'nohm:relations:',
+        scoredindex: 'nohm:scoredindex:',
+        unique: 'nohm:uniques:',
+      },
+      publish: false,
+    }
+
+> Snapshot 3
+
+    NohmClass {
+      LinkError: Function LinkError {},
+      ValidationError: Function ValidationError {},
+      extraValidators: [],
+      isPublishSubscribed: false,
+      meta: true,
+      modelCache: {},
+      prefix: {
+        channel: 'nohm:channel:',
+        hash: 'nohm:hash:',
+        idsets: 'nohm:idsets:',
+        incrementalIds: 'nohm:ids:',
+        index: 'nohm:index:',
+        meta: {
+          idGenerator: 'nohm:meta:idGenerator:',
+          properties: 'nohm:meta:properties:',
+          version: 'nohm:meta:version:',
+        },
+        relationKeys: 'nohm:relationKeys:',
+        relations: 'nohm:relations:',
+        scoredindex: 'nohm:scoredindex:',
+        unique: 'nohm:uniques:',
+      },
+      publish: false,
+    }
+
+> Snapshot 4
+
+    NohmClass {
+      LinkError: Function LinkError {},
+      ValidationError: Function ValidationError {},
+      extraValidators: [],
+      isPublishSubscribed: false,
+      meta: true,
+      modelCache: {},
+      prefix: {
+        channel: 'nohm:channel:',
+        hash: 'nohm:hash:',
+        idsets: 'nohm:idsets:',
+        incrementalIds: 'nohm:ids:',
+        index: 'nohm:index:',
+        meta: {
+          idGenerator: 'nohm:meta:idGenerator:',
+          properties: 'nohm:meta:properties:',
+          version: 'nohm:meta:version:',
+        },
+        relationKeys: 'nohm:relationKeys:',
+        relations: 'nohm:relations:',
+        scoredindex: 'nohm:scoredindex:',
+        unique: 'nohm:uniques:',
+      },
+      publish: false,
+    }
