+++ test/snapshots/features.test.ts.md
@@ -1,0 +1,26 @@
+# Snapshot report for `test/features.test.ts`
+
+The actual snapshot is saved in `features.test.ts.snap`.
+
+Generated by [AVA](https://ava.li).
+
+## setPrefix
+
+> Setting a custom prefix did not work as expected
+
+    {
+      channel: 'hurgel:channel:',
+      hash: 'hurgel:hash:',
+      idsets: 'hurgel:idsets:',
+      incrementalIds: 'hurgel:ids:',
+      index: 'hurgel:index:',
+      meta: {
+        idGenerator: 'hurgel:meta:idGenerator:',
+        properties: 'hurgel:meta:properties:',
+        version: 'hurgel:meta:version:',
+      },
+      relationKeys: 'hurgel:relationKeys:',
+      relations: 'hurgel:relations:',
+      scoredindex: 'hurgel:scoredindex:',
+      unique: 'hurgel:uniques:',
+    }
