+++ test/tslint.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../tslint.json",
+  "rules": {
+    "object-literal-sort-keys": [true, "match-declaration-order-only"]
+  }
+}
